feat(rng): Save API workflow
