+++ rucomfyui-node-graph/src/main.rs
@@ -18,9 +18,12 @@ use rucomfyui::{
         Object, ObjectInfo, ObjectInputMetaTyped, ObjectInputMetaTypedRoundValue, ObjectInputType,
         ObjectType,
     },
-    workflow::{WorkflowInput, WorkflowMeta, WorkflowNode, WorkflowNodeId},
-    Workflow, WorkflowGraph,
+    workflow::{WorkflowInput, WorkflowNode, WorkflowNodeId},
+    Workflow,
 };
+
+mod workflow;
+use workflow::{node_graph_to_workflow_graph, NodeToWorkflowNodeMapping};
 
 fn main() {
     eframe::run_native(
@@ -35,7 +38,6 @@ fn main() {
     .unwrap();
 }
 
-type NodeToWorkflowNodeMapping = HashMap<NodeId, WorkflowNodeId>;
 pub struct Application {
     persisted: PersistedState,
     user_state: FlowUserState,
@@ -50,8 +52,14 @@ pub struct Application {
     error: Option<(String, String)>,
 
     file_dialog: egui_file_dialog::FileDialog,
+    file_mode: FileMode,
 
     last_queue_time: Option<Instant>,
+}
+enum FileMode {
+    None,
+    Load,
+    Save,
 }
 impl Application {
     pub fn new(cc: &eframe::CreationContext<'_>) -> Self {
@@ -83,12 +91,13 @@ impl Application {
             error: None,
 
             file_dialog,
+            file_mode: FileMode::None,
 
             last_queue_time: None,
         }
     }
 
-    fn load(&mut self, path: PathBuf) -> anyhow::Result<()> {
+    fn api_load(&mut self, path: PathBuf) -> anyhow::Result<()> {
         let object_info = self.object_info.as_ref().context("No object info")?;
         let workflow = rucomfyui::Workflow::from_json(
             &std::fs::read_to_string(path).context("Failed to read file")?,
@@ -147,6 +156,18 @@ impl Application {
         Ok(())
     }
 
+    fn api_save(&self, path: PathBuf) -> anyhow::Result<()> {
+        let (workflow, _) = node_graph_to_workflow_graph(&self.persisted.state.graph);
+        std::fs::write(
+            path,
+            workflow
+                .into_workflow()
+                .to_json()
+                .context("Failed to serialize workflow")?,
+        )?;
+        Ok(())
+    }
+
     fn process_incoming_events(&mut self) -> bool {
         let mut needs_repaint = false;
 
@@ -198,10 +219,24 @@ impl Application {
 
         // Process file dialog
         if let Some(path) = self.file_dialog.take_selected() {
-            if let Err(err) = self.load(path) {
-                self.error = Some(("File load".to_string(), err.to_string()));
-                needs_repaint = true;
+            match self.file_mode {
+                FileMode::None => {
+                    eprintln!("File mode is None, this should not happen. Path: {path:?}");
+                }
+                FileMode::Load => {
+                    if let Err(err) = self.api_load(path) {
+                        self.error = Some(("File load".to_string(), err.to_string()));
+                        needs_repaint = true;
+                    }
+                }
+                FileMode::Save => {
+                    if let Err(err) = self.api_save(path) {
+                        self.error = Some(("File save".to_string(), err.to_string()));
+                        needs_repaint = true;
+                    }
+                }
             }
+            self.file_mode = FileMode::None;
         }
 
         needs_repaint
@@ -220,70 +255,13 @@ impl Application {
     }
 
     fn queue(&mut self) {
-        let g = WorkflowGraph::new();
-        let graph = &self.persisted.state.graph;
-        let mut mapping: HashMap<NodeId, WorkflowNodeId> = HashMap::new();
-        let mut input_mapping: HashMap<InputId, (WorkflowNodeId, String)> = HashMap::new();
-        let mut output_mapping: HashMap<OutputId, (WorkflowNodeId, usize)> = HashMap::new();
-
-        for (node_id, node) in &graph.nodes {
-            let object = &node.user_data.template.0;
-            let mut g_node = WorkflowNode::new(object.name.clone())
-                .with_meta(WorkflowMeta::new(object.display_name.clone()));
-
-            let mut connections = vec![];
-            for (input_name, input_id) in &node.inputs {
-                let input = graph.inputs.get(*input_id).unwrap();
-                let workflow_input = match &input.value {
-                    FlowValueType::Array { selected, .. } => {
-                        WorkflowInput::String(selected.clone())
-                    }
-                    FlowValueType::String { value, .. } => WorkflowInput::String(value.clone()),
-                    FlowValueType::Float { value, .. } => WorkflowInput::F64(*value),
-                    FlowValueType::SignedInt { value, .. } => WorkflowInput::I64(*value),
-                    FlowValueType::UnsignedInt { value, .. } => WorkflowInput::U64(*value),
-                    FlowValueType::Boolean(value) => WorkflowInput::Boolean(*value),
-                    FlowValueType::Other(_) => {
-                        connections.push((*input_id, input_name.clone()));
-                        continue;
-                    }
-                    FlowValueType::Unknown => continue,
-                };
-                g_node.add_input(input_name.clone(), workflow_input);
-            }
-
-            let g_node_id = g.add_dynamic(g_node);
-            mapping.insert(node_id, g_node_id);
-            for (input_id, input_name) in connections {
-                input_mapping.insert(input_id, (g_node_id, input_name));
-            }
-            for (output_slot, (_output_name, output_id)) in node.outputs.iter().enumerate() {
-                output_mapping.insert(*output_id, (g_node_id, output_slot));
-            }
-        }
-
-        for (input_id, output_ids) in &graph.connections {
-            let Some(output_id) = output_ids.first().copied() else {
-                continue;
-            };
-
-            let Some((g_input_node_id, input_name)) = input_mapping.get(&input_id) else {
-                continue;
-            };
-            let Some((g_output_node_id, output_slot)) = output_mapping.get(&output_id) else {
-                continue;
-            };
-            let Some(mut g_input_node) = g.get_node_mut(*g_input_node_id) else {
-                continue;
-            };
-            g_input_node.add_input(
-                input_name.clone(),
-                WorkflowInput::slot(*g_output_node_id, *output_slot as u32),
-            );
-        }
+        let (graph, mapping) = node_graph_to_workflow_graph(&self.persisted.state.graph);
 
         self.tokio_input_tx
-            .send(TokioInputEvent::QueueWorkflow((mapping, g.into_workflow())))
+            .send(TokioInputEvent::QueueWorkflow((
+                mapping,
+                graph.into_workflow(),
+            )))
             .unwrap();
         self.last_queue_time = Some(Instant::now());
     }
@@ -304,7 +282,13 @@ impl eframe::App for Application {
                 egui::widgets::global_dark_light_mode_switch(ui);
                 if is_connected {
                     if ui.button("Open API workflow").clicked() {
+                        self.file_mode = FileMode::Load;
                         self.file_dialog.select_file();
+                    }
+
+                    if ui.button("Save API workflow").clicked() {
+                        self.file_mode = FileMode::Save;
+                        self.file_dialog.save_file();
                     }
 
                     if let Some(last_queue_time) = self.last_queue_time {
